update of host manage script

+++ tools/run_on_rpi_all.sh
@@ -1,11 +1,24 @@
 #!/bin/bash
 
 script=$1
-KEY=.ssh/rpik8_rsa
+LOG_CMD=${2:-cat}
+KEY=$HOME/.ssh/rpik8_rsa
 ALL="rpi3-1 rpi4-1 rpi4-2 rpi4-3"
+
 for node in $ALL ; do
-   scp -i $KEY -l pi $script $node:
-   ssh -i $KEY -l pi $node ./$script
+    echo =========== $node
+    if [ "cat" != "$LOG_CMD" ] ; then
+        LOG_EXEC="tee $LOG_CMD.$node"
+    else
+        LOG_EXEC="cat"
+    fi
+
+    if [ -e "$script" ] ; then 
+        scp -i $KEY -l pi $script $node:
+        ssh -i $KEY -l pi $node ./$script | $LOG_EXEC
+    else
+        ssh -i $KEY -l pi $node "$script" | $LOG_EXEC
+    fi
 done
